DOC-8536 use agent scope
Brings (mostly) inline with Python app and spec.

Note that the `hotels` search inflates the results with a KV get
with `project` parameter, rather than the subdoc manipulations
specified. But the `project` version does seem to do very similar,
so will check with PM and engineering if this is equivalent.